update logic to detect if running in Alteryx
There are situations where the AlteryxRDataX package will need to be explicitly loaded for code to be run outside of Alteryx. The current detection mechanism breaks in this situation. The alternate approach used here uses one of the global variables introduced in the environment inside the R tool to detect if the code is being run inside or outside of Alteryx.